+++ tests/fields/test_passive_hidden_field.py
@@ -10,10 +10,10 @@ class TestPassiveHiddenField(FormTestCase):
         class MyForm(Form):
             id = PassiveHiddenField()
 
-        class A(object):
+        class A:
             id = None
 
-        form = MyForm(MultiDict({'id': 12}))
+        form = MyForm(MultiDict({"id": 12}))
         a = A()
         form.populate_obj(a)
         assert a.id is None