+++ tests/fields/test_select_multiple_field.py
@@ -5,22 +5,21 @@ from tests import MultiDict
 from wtforms_components import SelectMultipleField
 
 
-class Dummy(object):
+class Dummy:
     fruits = []
 
 
 class TestSelectMultipleField(FormTestCase):
     choices = (
-        ('Fruits', (
-            ('apple', 'Apple'),
-            ('peach', 'Peach'),
-            ('pear', 'Pear')
-        )),
-        ('Vegetables', (
-            ('cucumber', 'Cucumber'),
-            ('potato', 'Potato'),
-            ('tomato', 'Tomato'),
-        ))
+        ("Fruits", (("apple", "Apple"), ("peach", "Peach"), ("pear", "Pear"))),
+        (
+            "Vegetables",
+            (
+                ("cucumber", "Cucumber"),
+                ("potato", "Potato"),
+                ("tomato", "Tomato"),
+            ),
+        ),
     )
 
     def init_form(self, **kwargs):
@@ -32,27 +31,17 @@ class TestSelectMultipleField(FormTestCase):
 
     def test_understands_nested_choices(self):
         form_class = self.init_form(choices=self.choices)
-        form = form_class(
-            MultiDict([
-                ('fruits', 'apple'),
-                ('fruits', 'invalid')
-            ])
-        )
+        form = form_class(MultiDict([("fruits", "apple"), ("fruits", "invalid")]))
         form.validate()
 
         assert form.errors == {
-            'fruits': [u"'invalid' is not a valid choice for this field"]
+            "fruits": ["'invalid' is not a valid choice for this field"]
         }
 
     def test_option_selected(self):
         form_class = self.init_form(choices=self.choices)
 
         obj = Dummy()
-        obj.fruits = ['peach']
-        form = form_class(
-            obj=obj
-        )
-        assert (
-            '<option selected value="peach">Peach</option>' in
-            str(form.fruits)
-        )
+        obj.fruits = ["peach"]
+        form = form_class(obj=obj)
+        assert '<option selected value="peach">Peach</option>' in str(form.fruits)