+++ tests/test_chain_validator.py
@@ -1,10 +1,14 @@
 from wtforms import Form
-from wtforms.fields import TextField
 from wtforms.validators import DataRequired, Email
 from wtforms_test import FormTestCase
 
 from tests import MultiDict
 from wtforms_components import Chain
+
+try:
+    from wtforms.fields import TextField
+except ImportError:  # wtforms3
+    from wtforms.fields import StringField as TextField
 
 
 class TestChainValidator(FormTestCase):
@@ -12,6 +16,6 @@ class TestChainValidator(FormTestCase):
         class MyForm(Form):
             email = TextField(validators=[Chain([DataRequired(), Email()])])
 
-        form = MyForm(MultiDict({'name': ''}))
+        form = MyForm(MultiDict({"name": ""}))
         form.validate()
-        assert 'email' in form.errors
+        assert "email" in form.errors