+++ wtforms_components/__init__.py
@@ -15,18 +15,19 @@ from .fields import (
     IntegerField,
     IntegerSliderField,
     IntIntervalField,
+    JSONField,
     PassiveHiddenField,
     SearchField,
     SelectField,
     SelectMultipleField,
     SplitDateTimeField,
     StringField,
-    TimeField
+    TimeField,
 )
 from .validators import Chain, DateRange, Email, If, TimeRange
 from .widgets import NumberInput, ReadOnlyWidgetProxy, SelectWidget
 
-__version__ = '0.10.0'
+__version__ = "0.10.5"
 
 
 __all__ = (
@@ -48,6 +49,7 @@ __all__ = (
     IntegerField,
     IntegerSliderField,
     IntIntervalField,
+    JSONField,
     NumberInput,
     PassiveHiddenField,
     SearchField,
@@ -57,7 +59,7 @@ __all__ = (
     SplitDateTimeField,
     StringField,
     TimeField,
-    TimeRange
+    TimeRange,
 )
 
 
@@ -65,10 +67,9 @@ class ModelForm(Form):
     """
     Simple ModelForm, use this if your form needs to use the Unique validator
     """
-    def __init__(self, formdata=None, obj=None, prefix='', **kwargs):
-        Form.__init__(
-            self, formdata=formdata, obj=obj, prefix=prefix, **kwargs
-        )
+
+    def __init__(self, formdata=None, obj=None, prefix="", **kwargs):
+        Form.__init__(self, formdata=formdata, obj=obj, prefix=prefix, **kwargs)
         self._obj = obj