+++ wtforms_components/fields/__init__.py
@@ -10,15 +10,16 @@ from .html5 import (
     IntegerField,
     IntegerSliderField,
     SearchField,
-    StringField
+    StringField,
 )
 from .interval import (
     DateIntervalField,
     DateTimeIntervalField,
     DecimalIntervalField,
     FloatIntervalField,
-    IntIntervalField
+    IntIntervalField,
 )
+from .json_field import JSONField
 from .passive_hidden import PassiveHiddenField
 from .select import SelectField
 from .select_multiple import SelectMultipleField
@@ -41,6 +42,7 @@ __all__ = (
     IntegerField,
     IntegerSliderField,
     IntIntervalField,
+    JSONField,
     PassiveHiddenField,
     SearchField,
     SelectField,