+++ wtforms_components/fields/html5.py
@@ -1,5 +1,24 @@
+try:
+    from wtforms.fields.html5 import (
+        DateField,
+        DateTimeField,
+        DecimalField,
+        DecimalRangeField,
+        IntegerField,
+        IntegerRangeField,
+        SearchField,
+    )
+except ImportError:  # wtforms>=3
+    from wtforms.fields import (
+        DateField,
+        DateTimeField,
+        DecimalField,
+        DecimalRangeField,
+        IntegerField,
+        IntegerRangeField,
+        SearchField,
+    )
 from wtforms.fields import StringField as _StringField
-from wtforms.fields import html5
 
 from ..widgets import (
     DateInput,
@@ -9,7 +28,7 @@ from ..widgets import (
     NumberInput,
     RangeInput,
     SearchInput,
-    TextInput
+    TextInput,
 )
 
 
@@ -17,48 +36,40 @@ class EmailField(_StringField):
     widget = EmailInput()
 
 
-class IntegerField(html5.IntegerField):
-    widget = NumberInput(step='1')
+class IntegerField(IntegerField):
+    widget = NumberInput(step="1")
 
 
-class DecimalField(html5.DecimalField):
-    widget = NumberInput(step='any')
+class DecimalField(DecimalField):
+    widget = NumberInput(step="any")
 
 
-class DateTimeLocalField(html5.DateTimeField):
+class DateTimeLocalField(DateTimeField):
     def __init__(
-        self,
-        label=None,
-        validators=None,
-        format='%Y-%m-%dT%H:%M:%S',
-        **kwargs
+        self, label=None, validators=None, format="%Y-%m-%dT%H:%M:%S", **kwargs
     ):
-        super(DateTimeLocalField, self).__init__(
-            label,
-            validators,
-            format,
-            **kwargs
-        )
+        super().__init__(label, validators, format, **kwargs)
+
     widget = DateTimeLocalInput()
 
 
-class DateTimeField(html5.DateTimeField):
+class DateTimeField(DateTimeField):
     widget = DateTimeInput()
 
 
-class DateField(html5.DateField):
+class DateField(DateField):
     widget = DateInput()
 
 
-class IntegerSliderField(html5.IntegerRangeField):
-    widget = RangeInput(step='1')
+class IntegerSliderField(IntegerRangeField):
+    widget = RangeInput(step="1")
 
 
-class DecimalSliderField(html5.DecimalRangeField):
-    widget = RangeInput(step='any')
+class DecimalSliderField(DecimalRangeField):
+    widget = RangeInput(step="any")
 
 
-class SearchField(html5.SearchField):
+class SearchField(SearchField):
     widget = SearchInput()