+++ wtforms_components/fields/passive_hidden.py
@@ -5,5 +5,6 @@ class PassiveHiddenField(HiddenField):
     """
     HiddenField that does not populate obj values.
     """
+
     def populate_obj(self, obj, name):
         pass