+++ wtforms_components/widgets.py
@@ -1,10 +1,9 @@
-from cgi import escape
 from copy import copy
 
-import six
+from markupsafe import Markup, escape
 from wtforms.validators import DataRequired, NumberRange
+from wtforms.widgets import Input, html_params
 from wtforms.widgets import Select as _Select
-from wtforms.widgets import html_params, HTMLString, Input
 
 from .validators import DateRange, TimeRange
 
@@ -41,9 +40,9 @@ def min_max(field, validator_class):
 
     data = {}
     if min_values:
-        data['min'] = max(min_values)
+        data["min"] = max(min_values)
     if max_values:
-        data['max'] = min(max_values)
+        data["max"] = min(max_values)
     return data
 
 
@@ -55,30 +54,31 @@ def has_validator(field, validator_class):
     :param field: WTForms Field object
     :param validator_class: WTForms Validator class
     """
-    return any([
-        isinstance(validator, validator_class)
-        for validator in field.validators
-    ])
+    return any(
+        [isinstance(validator, validator_class) for validator in field.validators]
+    )
 
 
 class HTML5Input(Input):
+    validation_attrs = ["required", "disabled"]
+
     def __init__(self, **kwargs):
         self.options = kwargs
 
     def __call__(self, field, **kwargs):
         if has_validator(field, DataRequired):
-            kwargs.setdefault('required', True)
+            kwargs.setdefault("required", True)
 
         for key, value in self.range_validators(field).items():
             kwargs.setdefault(key, value)
 
-        if hasattr(field, 'widget_options'):
+        if hasattr(field, "widget_options"):
             for key, value in self.field.widget_options:
                 kwargs.setdefault(key, value)
 
         options_copy = copy(self.options)
         options_copy.update(kwargs)
-        return super(HTML5Input, self).__call__(field, **options_copy)
+        return super().__call__(field, **options_copy)
 
     def range_validators(self, field):
         return {}
@@ -89,75 +89,84 @@ class BaseDateTimeInput(HTML5Input):
     Base class for TimeInput, DateTimeLocalInput, DateTimeInput and
     DateInput widgets
     """
+
     range_validator_class = DateRange
 
     def range_validators(self, field):
         data = min_max(field, self.range_validator_class)
-        if 'min' in data:
-            data['min'] = data['min'].strftime(self.format)
-        if 'max' in data:
-            data['max'] = data['max'].strftime(self.format)
+        if "min" in data:
+            data["min"] = data["min"].strftime(self.format)
+        if "max" in data:
+            data["max"] = data["max"].strftime(self.format)
         return data
 
 
 class TextInput(HTML5Input):
-    input_type = 'text'
+    input_type = "text"
 
 
 class SearchInput(HTML5Input):
     """
     Renders an input with type "search".
     """
-    input_type = 'search'
+
+    input_type = "search"
 
 
 class MonthInput(HTML5Input):
     """
     Renders an input with type "month".
     """
-    input_type = 'month'
+
+    input_type = "month"
 
 
 class WeekInput(HTML5Input):
     """
     Renders an input with type "week".
     """
-    input_type = 'week'
+
+    input_type = "week"
 
 
 class RangeInput(HTML5Input):
     """
     Renders an input with type "range".
     """
-    input_type = 'range'
+
+    input_type = "range"
 
 
 class URLInput(HTML5Input):
     """
     Renders an input with type "url".
     """
-    input_type = 'url'
+
+    input_type = "url"
 
 
 class ColorInput(HTML5Input):
     """
-    Renders an input with type "tel".
+    Renders an input with type "color".
     """
-    input_type = 'color'
+
+    input_type = "color"
 
 
 class TelInput(HTML5Input):
     """
     Renders an input with type "tel".
     """
-    input_type = 'tel'
+
+    input_type = "tel"
 
 
 class EmailInput(HTML5Input):
     """
     Renders an input with type "email".
     """
-    input_type = 'email'
+
+    input_type = "email"
 
 
 class TimeInput(BaseDateTimeInput):
@@ -167,9 +176,10 @@ class TimeInput(BaseDateTimeInput):
     Adds min and max html5 field parameters based on field's TimeRange
     validator.
     """
-    input_type = 'time'
+
+    input_type = "time"
     range_validator_class = TimeRange
-    format = '%H:%M:%S'
+    format = "%H:%M:%S"
 
 
 class DateTimeLocalInput(BaseDateTimeInput):
@@ -179,8 +189,9 @@ class DateTimeLocalInput(BaseDateTimeInput):
     Adds min and max html5 field parameters based on field's DateRange
     validator.
     """
-    input_type = 'datetime-local'
-    format = '%Y-%m-%dT%H:%M:%S'
+
+    input_type = "datetime-local"
+    format = "%Y-%m-%dT%H:%M:%S"
 
 
 class DateTimeInput(BaseDateTimeInput):
@@ -190,8 +201,9 @@ class DateTimeInput(BaseDateTimeInput):
     Adds min and max html5 field parameters based on field's DateRange
     validator.
     """
-    input_type = 'datetime'
-    format = '%Y-%m-%dT%H:%M:%SZ'
+
+    input_type = "datetime"
+    format = "%Y-%m-%dT%H:%M:%SZ"
 
 
 class DateInput(BaseDateTimeInput):
@@ -201,8 +213,9 @@ class DateInput(BaseDateTimeInput):
     Adds min and max html5 field parameters based on field's DateRange
     validator.
     """
-    input_type = 'date'
-    format = '%Y-%m-%d'
+
+    input_type = "date"
+    format = "%Y-%m-%d"
 
 
 class NumberInput(HTML5Input):
@@ -212,14 +225,15 @@ class NumberInput(HTML5Input):
     Adds min and max html5 field parameters based on field's NumberRange
     validator.
     """
-    input_type = 'number'
+
+    input_type = "number"
     range_validator_class = NumberRange
 
     def range_validators(self, field):
         return min_max(field, self.range_validator_class)
 
 
-class ReadOnlyWidgetProxy(object):
+class ReadOnlyWidgetProxy:
     def __init__(self, widget):
         self.widget = widget
 
@@ -227,7 +241,10 @@ class ReadOnlyWidgetProxy(object):
         return getattr(self.widget, name)
 
     def __call__(self, field, **kwargs):
-        kwargs.setdefault('readonly', True)
+        kwargs.setdefault("readonly", True)
+        # Some html elements also need disabled attribute to achieve the
+        # expected UI behaviour.
+        kwargs.setdefault("disabled", True)
         return self.widget(field, **kwargs)
 
 
@@ -235,6 +252,7 @@ class SelectWidget(_Select):
     """
     Add support of choices with ``optgroup`` to the ``Select`` widget.
     """
+
     @classmethod
     def render_optgroup(cls, value, label, mixed):
         children = []
@@ -243,9 +261,9 @@ class SelectWidget(_Select):
             item_html = cls.render_option(item_value, item_label, mixed)
             children.append(item_html)
 
-        html = u'<optgroup label="%s">%s</optgroup>'
-        data = (escape(six.text_type(value)), u'\n'.join(children))
-        return HTMLString(html % data)
+        html = '<optgroup label="%s">%s</optgroup>'
+        data = (escape(str(value)), "\n".join(children))
+        return Markup(html % data)
 
     @classmethod
     def render_option(cls, value, label, mixed):
@@ -256,18 +274,22 @@ class SelectWidget(_Select):
         if isinstance(label, (list, tuple)):
             return cls.render_optgroup(value, label, mixed)
 
-        coerce_func, data = mixed
-        if isinstance(data, list) or isinstance(data, tuple):
-            selected = coerce_func(value) in data
+        try:
+            coerce_func, data = mixed
+        except TypeError:
+            selected = mixed
         else:
-            selected = coerce_func(value) == data
+            if isinstance(data, list) or isinstance(data, tuple):
+                selected = coerce_func(value) in data
+            else:
+                selected = coerce_func(value) == data
 
-        options = {'value': value}
+        options = {"value": value}
 
         if selected:
-            options['selected'] = True
+            options["selected"] = True
 
-        html = u'<option %s>%s</option>'
-        data = (html_params(**options), escape(six.text_type(label)))
+        html = "<option %s>%s</option>"
+        data = (html_params(**options), escape(str(label)))
 
-        return HTMLString(html % data)
+        return Markup(html % data)